refactor(app): extract useProjects hook (git polling + bridge IPC)
Phase 1 step 5 of the god-file cleanup plan. Move project-scoped I/O
out of App.tsx into src/hooks/useProjects.ts:

- gitStatusRef (cache of git_status results keyed by absolute path).
- gitPollingRef (in-flight guard for the 30s poller).
- refreshGitStatusFor / refreshAllGitStatus.
- The 30s + window-focus git poll useEffect.
- announceProjectToBridge / watchProjectForBridge /
  unwatchProjectForBridge (Tauri invokes for cwd hand-off and
  project-extension hot-reload watching).
- GitStatus type.

Hook accepts ctx callbacks: getProjectPaths() so the poller always
sees the live list, and onGitStatusChanged() wired to
syncProjectsToState so /sidebar/projects badges refresh.

Project list state itself (projectsRef, projectsLoadedRef,
tabBucketsRef) plus orchestrators (openProjectByPath et al.) stay
inline because they're entangled with tab-bucket swap and terminal
replay — those move when useTabs lands.

App.tsx 5763 -> 5678 lines. tsc clean, eslint clean, 305/305 tests
pass. No behavior change.
